requirements.txt
solara
plaid-python
python-dotenv
firebase-admin
pandas
plotly
authlib
requests
